+++ packages/taro-quickapp/src/lifecycle.js
@@ -201,6 +201,7 @@ function doUpdate (component, prevProps, prevState) {
     invokeEffects(component)
     if (component['$$hasLoopRef']) {
       Current.current = component
+      Current.index = 0
       component._disableEffect = true
       component._createData(component.state, component.props, true)
       component._disableEffect = false