+++ packages/taro/src/index.js
@@ -8,7 +8,7 @@ import { getOriginal as internal_get_original } from './internal/get-original'
 import { getEnv, ENV_TYPE } from './env'
 import Events from './events'
 import render from './render'
-import { createRef, commitAttachRef, detachAllRef, RefsArray } from './ref'
+import { createRef, commitAttachRef, detachAllRef, RefsArray, handleLoopRef } from './ref'
 import Link from './interceptor'
 import * as interceptors from './interceptor/interceptors'
 import {
@@ -76,6 +76,7 @@ export {
   Link,
   interceptors,
   RefsArray,
+  handleLoopRef,
   Current,
   useEffect,
   useLayoutEffect,
@@ -123,6 +124,7 @@ export default {
   Link,
   interceptors,
   RefsArray,
+  handleLoopRef,
   Current,
   useEffect,
   useLayoutEffect,